requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.5.1
pydantic==2.8.2
python-dotenv==1.0.1
PyJWT==2.9.0
pymongo==4.7.2
joblib==1.4.2
numpy==2.1.3
APScheduler==3.10.4
